FWStructureにSetPose()を追加
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@4529 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Framework/FWScene.cpp
@@ -237,7 +237,7 @@ FWStructureIf* FWScene::GetFWStructure(){
 }
 
 FWStructureIf* FWScene::GetFWStructure(int n){
-	int size = fwStructures.size();
+	int size = (int)fwStructures.size();
 	if(n < size){
 		return fwStructures[n];
 	}
